backend/requirements.txt
sqlalchemy
aioredis
redis
tornado
pillow
bcrypt
psycopg2